requirements.txt
# Core ML/AI Libraries
torch==2.7.1
torchvision==0.22.1
torchaudio==2.5.1
numpy==1.26.4
scipy==1.16.0
scikit-learn==1.7.1

# Computer Vision & Object Detection
ultralytics==8.0.231
opencv-python==4.10.0
pillow==11.3.0

# Deep Learning & Transformers
transformers==4.41.1
sentence-transformers==2.6.1
huggingface-hub==0.33.4
tokenizers==0.19.1
safetensors==0.5.3

# Audio Processing
git+https://github.com/openai/whisper.git
openai==1.30.1
SpeechRecognition==3.14.3
pyaudio==0.2.14
pyttsx3==2.90
sounddevice==0.5.2
soundfile==0.13.1

# Vector Search & Embeddings
faiss==1.9.0
open_clip_torch==2.32.0

# Web Framework & API
fastapi==0.112.2
uvicorn==0.35.0
starlette==0.38.2

# Database
pymongo==4.7.2

# Hyperparameter Tuning
optuna==4.4.0

# Data Processing & Visualization
pandas==2.3.1
matplotlib==3.10.3
seaborn==0.13.2
plotly==6.2.0

# Utilities
requests==2.32.4
tqdm==4.67.1
pyyaml==6.0.2
click==8.1.8

# Windows-specific
pywin32==308
pipwin==0.5.2

# Development & Jupyter (optional)
ipython==9.4.0
jupyter_client==8.6.3